+++ NEWS.md
@@ -14,6 +14,7 @@
 * Fix basisOfRecord reference in `convert_to_dwca()`.
 * Allow only one basisOfRecord `create_eml()`.
 * Remove XML attributes to prevent id clashing and schema invalidation when constructing provenance nodes in `create_eml()`.
+* Remove all L0 data entities from the L1 EML, these should not be inherited by the L1 in `create_eml()`.
 
 # ecocomDP 1.0.0